+++ execution/playground/playground_test.go
@@ -1,16 +1,10 @@
 package playground
 
 import (
-	"bytes"
-	"os"
 	"testing"
 
-	"github.com/davecgh/go-spew/spew"
-	"github.com/jensneuse/diffview"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/wundergraph/graphql-go-tools/v2/pkg/testing/goldie"
 )
 
 func TestNew(t *testing.T) {
@@ -49,18 +43,7 @@ func TestConfigureHandlers(t *testing.T) {
 			handlers[i].Handler = nil
 		}
 
-		var out bytes.Buffer
-		spew.Fdump(&out, handlers)
-
-		goldie.Assert(t, "handlers", out.Bytes())
-		if t.Failed() {
-			fixture, err := os.ReadFile("./fixtures/handlers.golden")
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			diffview.NewGoland().DiffViewBytes("handlers", fixture, out.Bytes())
-		}
+		assert.Equal(t, "/playground", handlers[0].Path)
 	})
 
 	t.Run("should respect trailing slash for playground path", func(t *testing.T) {
